% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondFunds.R
\name{BondFunds}
\alias{BondFunds}
\title{Bond Funds}
\format{A data frame with 9 columns and 180 entries.}
\source{
Levine, David, David Stephan, and Kathryn Szabat.  Statistics for Managers Using Microsoft EXCEL, Seventh Edition.  New York: Pearson, 2014.
}
\description{
Select qualities of 180 bond funds.
\itemize{
\item Fund Number
\item Type
\item Assets
\item Fees
\item Expense Ratio
\item Return 2008
\item 3-Year Return
\item 5-Year Return
\item Risk
}
}

